% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElder.R
\name{addElder}
\alias{addElder}
\title{Plots the elder ray similar to the volume chart (histograms with green&red colours or defined by up.col & down.col)}
\usage{
addElder(d, log.scale = FALSE)
}
\arguments{
\item{log.scale}{- scaling parm taken as is from addVo}

\item{m}{- The two columns, elderH & elderL.}
}
\value{
same as addVo
}
\description{
The function returns the elders Ray for a given OHLCV
}
\author{
Siva Sunku
}
\keyword{Ray}
\keyword{elders}
