% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar}
\alias{bar}
\alias{is.bar}
\alias{is.pricehit.bar}
\alias{is.pricexed.bar}
\alias{isdoji.bar}
\alias{isgreen.bar}
\alias{isred.bar}
\title{Bars - Functions related to all the bars/candles}
\usage{
bar(b)

is.bar(b)

isgreen.bar(b, pct = 0)

isred.bar(b, pct = 0)

isdoji.bar(b, pct = 0)

is.pricehit.bar(b, price)

is.pricexed.bar(b, price, dir = 1)
}
\arguments{
\item{b}{- bar/OHLC,}

\item{pct}{- percentage of points considered in calculation}

\item{price}{- price}

\item{dir}{- direction (GREEN/LONG stands for 1, SHORT/RED stands for -1)}
}
\value{
TRUE/FALSE

isgreen.bar - returns TRUE/FALSE

isred.bar - returns TRUE/FALSE

isdoji.bar - returns TRUE/FALSE

is.pricehit.bar  - returns TRUE/FALSE

is.pricexed.bar  - returns TRUE/FALSE
}
\description{
Functions related to all the bars/candles
}
\details{
bar - converts OHLC object into bar class. Both are aliases infact

is.bar - checks if given object is a bar

isgreen.bar - evaluates if close of a bar is atleast percent greater than open

isred.bar - evaluates if close of a bar is atleast percent lessr than open

isdoji.bar - evaluates if close and open are near to each other by pct

is.pricehit.bar - evaluates if the price 'p' is hit by bar 'b'

is.pricexed.bar - evaluates if the bar 'b' closed on price 'p' in given direction 'dir'.
if direction is - LONG/GREEN(1) - b closes > p , returns TRUE
}
\note{

}
\author{
Siva Sunku
}
\keyword{bars}

