% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencyRatio.R
\name{efficiencyRatio}
\alias{efficiencyRatio}
\title{Calculate the efficiencyRatio}
\usage{
efficiencyRatio(m)
}
\arguments{
\item{m}{- input numbers}
}
\value{
returns the efficiency Ratio
}
\description{
efficiency Ratio is a trend strength measure

It is calculated by net displacement/abs(each displacement) in a given period
}
\details{
The function returns the efficiency Ratio for a given n-points
}
\examples{
efficiencyRatio(m)
}
\author{
Siva Sunku
}
\keyword{Ratio}
\keyword{efficiency}
