% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elasticity}
\alias{elasticity}
\title{Calculate the elders Ray of given OHLCV}
\usage{
elasticity(m, FUN = Cl)
}
\arguments{
\item{m}{- OHLC}

\item{FUN}{- Which price is OHLC to be used for calculating. Default Cl. If 
If Hi to be used, use FUN - Hi}
}
\value{
returns xts object with the elasticity
}
\description{
elders Ray gives two units bullPower, bearPower.

bullPower - Difference from High to EMA(n)

bearPower - Difference from Low to EMA(n)
}
\details{
The function returns the elders Ray for a given OHLCV
}
\examples{
elderRay(m,n=13)
}
\author{
Siva Sunku
}
\keyword{Ray}
\keyword{elders}

