% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elderRay.R
\name{elderRay}
\alias{elderRay}
\title{Calculate the elders Ray of given OHLCV}
\usage{
elderRay(m, n = 13)
}
\arguments{
\item{m}{- OHLC}

\item{n}{- number of units for calculating EMA, (default - 13)}
}
\value{
returns slope
}
\description{
elders Ray gives two units bullPower, bearPower.

bullPower - Difference from High to EMA(n)

bearPower - Difference from Low to EMA(n)
}
\details{
The function returns the elders Ray for a given OHLCV
}
\examples{
elderRay(m,n=13)
}
\author{
Siva Sunku
}
\keyword{Ray}
\keyword{elders}

