% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowDex.R
\name{flow}
\alias{flow}
\title{Calculate the flow and returns the}
\usage{
flow(m)
}
\arguments{
\item{m}{- OHLC}
}
\value{
returns the bearishFlow,bullishFlow
}
\description{
The function returns the difference of Highs and lows
If Candle is Green (Closed above previous Close) - Difference is - ( Close minus trueLow)
If Candle is Red   (Closed below previous Close) - Difference is - ( trueHigh minus Close)
trueHigh - Max of CurrentHigh,PrevClose
trueLow  - Min of CurrentLow, prevClose
}
\author{
Siva Sunku
}
\keyword{flowDex}
