% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowDex.R
\name{flowDex}
\alias{flowDex}
\title{Calculates the sum/mean of flow and returns the value}
\usage{
flowDex(m, fFUN = mean)
}
\arguments{
\item{m}{- xts}

\item{fFUN}{- function to be applied. like sum/mean}
}
\value{
returns the bearishFlow,bullishFlow
}
\description{
Calculates the sum/mean of flow and returns the value
}
\author{
Siva Sunku
}
\keyword{flowDex}
