% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{pivots}
\alias{pivots}
\title{Calculate the various pivots}
\usage{
pivots(m, type = c("regular", "woodies", "camarilla"))
}
\arguments{
\item{m}{- the OHLC xts object}

\item{type}{- type of the pivots to be calculated. default regular pivots.}
}
\value{
returns OHLC xts with pivots. If nothing matches to type, returns the xts object itself.
}
\description{
The function returns the xts object with r1..s3 pivots.
}
\examples{
pivots(xts, type="woodies")
}
\author{
Siva Sunku
}
\keyword{pivots}

