% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revApply.R
\name{revapply}
\alias{revapply}
\title{This will apply higher interval values to lower inverval datasets.}
\usage{
revapply(l, h, parms = NULL)
}
\arguments{
\item{l}{- lower ordered xts}

\item{h}{- higher ordered xts}

\item{parms}{- which all the columns to be applied from higher order to lower order. Default All}
}
\value{
higher ordered xts distributed across to all lower ordered ones. only index of lowerorder is used
}
\description{
- Apply higher interval values to lower values. e.g. If weekly indicator need to be applied to all daily rows for that
particular week.
}
