% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Calculate the slope of given n points}
\usage{
slope(n, pct = FALSE)
}
\arguments{
\item{n}{- input numbers}

\item{pct}{- If 'n' should be normalized before finding the slope.}
}
\value{
returns slope
}
\description{
The function returns the slope of given n points
}
\examples{
slope(n=c(10,20,30))
}
\author{
Siva Sunku
}
\keyword{slope}

