% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elderRay.R, R/elderRayM.R
\name{elderRay}
\alias{elderRay}
\alias{elderRay}
\title{Calculate the elders Ray of given OHLCV}
\usage{
elderRay(m, n = 13)

elderRay(m, n = 13)
}
\arguments{
\item{m}{- OHLC}

\item{n}{- number of units for calculating EMA, (default - 13)}

\item{FUN}{- function to be applied for averges. Default EMA}

\item{m}{- OHLC}

\item{eman}{- n for calculating EMA}

\item{pown}{- n for calculating the bull/bear}
}
\value{
returns slope

po
}
\description{
elders Ray gives two units bullPower, bearPower.

bullPower - Difference from High to EMA(n)

bearPower - Difference from Low to EMA(n)

elders Ray gives two units bullPower, bearPower.

bullPower - Difference from High to EMA(n)

bearPower - Difference from Low to EMA(n)
}
\details{
The function returns the elders Ray for a given OHLCV

The function returns the elders Ray for a given OHLCV
}
\examples{
elderRay(m,n=13)
elderRay(m,n=13)
}
\author{
Siva Sunku

Siva Sunku
}
\keyword{Ray}
\keyword{elders}
